Implemented InputConfigurationManager and InputConfigurationBuilder classes
This commit provides mechanisms for InputConfigurationSpecificationDialog to manage the specification it creates. InputConfigurationManager will serve as a controller for the dialog, while InputConfigurationBuilder will provide persistence for the InputConfiguration managed by InputConfigurationManager. These mechanisms are a necessary step to decoupling this simulation specification logic in the workbench, from the data used to execute a simulation.