Added additional provenance support for runRemoteScript in the script manager.
This commit is another step towards allowing query support to be expanded. Inference cannot be undertaken without all of the steps of script execution (such as specifying local files, showing them being copied to the server, showing them being copied to different folders). Yes, the input files origin may be used for the execution description, but it won't make sense if execution happens on a different machine then were we say the files originated without showing that they have moved.