Several updates are included, in order to support evaluation for eScience 2015 publication.
These updates include: 

Local-execution bug fixes: Mostly for Mac OS-X. Modifications exclusive to ScriptManager. Path names within the generated scripts have been adapted to absolute path estimations for posix systems, rather than references to symbolic links to the user's home directory (e.g. ~/).

Build option: The Script Specification dialog now has a drop down for whether or not to clean and build the simulator prior to execution. This not only changes the dialog, but adds to the simulation specification object, and project (elements in the XML as well as getters and setters).

Performance Evaluation Mechanisms: Performance of GUI actions with and without provenance are now logged, along with the elapsed time within provenance enabled conditionals.

Script Output Filenames: Simulator std out redirects to a file that includes the name of the project and script version, to allow for multiple concurrent simulation executions within the same BrainGrid folder. This is also the case for the script execution log (which is used to recover provenance after a simulation). In addition, this is also the case for the SHA1 key redirect file (created through git log > proj_ver_SHA1Key.txt). Finally, the script name has been changed to also contain the name of the project, as well as the version. All of these changes have necessitated changing the logic for downloading the output files, as well.

There may be additional changes in this commit. Consider this as a baseline for changes after the merge with the main branch.