(1) Added simulation configuration test model. (2) Changed look and feel of SimulationConfigurationDialog. (3) Fixed model load bug in SimulationConfigurationDialog. (4) Commented IDE errors out of new project manager.
(1) The test model for Aaron's simulations configuration dialog has been added to the root of the workbench folder in the repository. The main method stub of the SimulationConfigurationDialog still points to an absolute location. This string just needs to be changed to the location of SimulationConfigurationTest.xml previously mentioned (hard-coded, but this main method is only a stub. The live version will pop a file selection dialog for the user to select the model they want to import)
(2) The Nimbus look and feel matches the rest of the workbench.
(3) Bugfix: Changed code in the main method stub of SimulationConfigurationTest to load XML doc from the test model file prior to instantiation of the SimulationConfigurationDialog.
(4) These errors are in a class that is under construction and not tied to the current workbench functionality. The remaining code was commented out to prevent the IDE from reporting the error and distracting from development.