Work-around for issue #137
Illegal memory access (out of bound) at the device function
changeDSSynapsePSR causes the crashes. It seems that iSync (index of
synapses) values may exceed the limit. I added the assertion to  check
the index value, but it doesn’t hit the assertion. Also the assertion
magically fix the problem. Similar thing happened when I added printf
statement in the device function.
I run growth_cuda w/test-small-connected.xml on cssgpu01 and cssgpu02p
and got the identical results.
This is a work-around of the issue, and we need to figure out the cause
of this problem.