Added AllSynapsesPolyFuncs_d.cu file
All global and device functions related to advanceSynapseDevice are
moved to AllSynapsesPolyFuncs_d.cu file.
Added setSynapseClassID function to set synapse class ID defined by
enumClassSynapses for the caller's Synapse class. The class ID will be
set to classSynapses_d in device memory, and the classSynapses_d will
be referred to call a device function for the particular synapse class.
Because we cannot use virtual function (Polymorphism) in device
functions, we use this scheme.