Added ClusterInfo class
All cluster specific information are moved from SimulationInfo into the
ClusterInfo.
The cluster contains a group of neurons and synapses that are regarded
as an execution unit. A simulation will include multiple clusters and
each cluster runs concurrently. A cluster will be assigned to a thread,
a GPU, or a computer node depending on a configuration.