remove tinyxml directory
TinyXPath has a dependency on TinyXML, and the necessary TinyXML
source code was provided with the raw download of TinyXPath. Existing
ParameterManager tests only use TinyXPath source files in compilation
and testing, so I am removing the /tinyxml files.

The TinyXPath downnload does not include the /docs subdirectory,
but in my humble opinion having the HTML documentation files in our
source code is not very important. What we should be doing (related
to #209) is including a documentation file that includes package names
and links to documentation for hard dependencies.