Use regular expressions to validate XPath results
Because TinyXPath does not verify that the retrieved value is
conventionally correct for the desired type, a secondary check
should be performed to ensure that any value conversion or
casting can be legally done and will prevent precision loss
(where applicable).

This commit implements regular expressions as the validation
mechanism, as opposed to substring checking.