Giving up right now on cleaning this up
There are just so many interdependencies between this class and other classes that making almost any changes breaks things, either causing growth to not compile (which is only moderately bad) or causing it to segfault (and good luck figuring out why it compiled but crashed). Here is a commit method from the branch containing the failed attempt at a more comprehensive cleanup:

Really, this is just a mess. Some simulation results are generated by recorder objects. Some are generated by "serialize" methods, which really don't serialize simulation state. There are member variables just in the connection classes that shouldn't be member variables, as they are merely used to hold temporary results in the middle of computations that are split across multiple private methods. And there are member variables that are accessed, directly or indirectly, by a rat's nest of methods in other classes. So, this compiles, but it segfaults at simulation start-up.