Fixed shift count over flow in EventQueue
This problem was first introduced at the commitment on Nov. 7th 2018 (633c40462d09dd148cb12a8845b426bf1681f8cb).
At the commitment the length of event queue was changed from uint32_t to uint64_t.

	Core/EventQueue.h

		#define BGQUEUE_ELEMENT uint32_t
—>
		#define BGQUEUE_ELEMENT uint64_t

Because of the change, the following statement doesn’t work correctly if idxQueue is greater than 32 (shift count overflow).

	queue & (0x1 << idxQueue)

So I fixed the statement as below and it now works.

	queue & (BGQUEUE_ELEMENT (0x1) << idxQueue)