Third-level Python wrapper
Run one second simulation with a single LIF neuron, pour regular input spike train, and record the psr of the synapse and the membrane voltage (vm) of the neuron.

To enable to record the psr and vm of the neuron for regular input spike trains, some changes are made on input module.